LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.isPointInStroke.scaleddashes-expected.txt

2d.path.isPointInStroke.scaleddashes
isPointInStroke() should return correct results on dashed paths at high scale factors
Actual output:

FAIL isPointInStroke() should return correct results on dashed paths at high scale factors assert_equals: ctx.isPointInStroke(3.74*scale, 14.46*scale) === false (got true[boolean], expected false[boolean]) expected false but got true
